python/requirements.txt
python-magic
pyyaml
typer
